// src/models/Mail.model.ts
export interface Mail {
    nomEntreprise?: string;
    contactEntreprise?: string;
    secteurActivite?: string;
    tailleEntreprise?: string;
    service?: string[];
    dateDemoCotation?: string;
    heureDemoCotation?: string;
    email: string;
    nom: string;
    prenoms: string;
    contactuser: string;
    message: string;
}